chat - setup and quotas tweaks
