+++ src/vs/workbench/workbench.desktop.main.ts
@@ -171,7 +171,6 @@ import './contrib/remoteTunnel/electron-sandbox/remoteTunnel.contribution.js';
 // Chat
 import './contrib/chat/electron-sandbox/chat.contribution.js';
 import './contrib/inlineChat/electron-sandbox/inlineChat.contribution.js';
-import './contrib/chat/browser/chatSetup.contribution.js';
 
 // Encryption
 import './contrib/encryption/electron-sandbox/encryption.contribution.js';
